remove old plasma coil
